review: change SafeContext test layout
